+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   - `R` - Frame and Filled Rect (cycled);
   - `S` - Selection.
 - Add: Files buttons can be reordered by drag-n-drop.
+- Add: Export current level body (or selection) as PNG image to a file or to
+  clipboard.
 
 ### sple.me